Make `FeedImageDataLoaderPresentationAdapter` to separate file
